hw04/requirements.txt
openai-whisper
torch
numpy